ui: Extract instance list component from home
Avoid mixing Vue template in Tornado Template.
