+++ ui/temboardui/static/src/components/home/Sparkline.vue
