+++ ui/temboardui/templates/home.html
@@ -20,70 +20,5 @@ var groups = {% raw json.dumps([i[0] for i in groups]) %};
   </a>
 </div>
 <div id="instances">
-  <div class="row">
-    <div class="col mb-2">
-      <form class="form-inline" onsubmit="event.preventDefault();">
-        <input type="text" class="form-control mr-sm-2" placeholder="Search instances" v-model="search">
-        <div class="dropdown mr-sm-2">
-          <button type="button" class="btn btn-secondary dropdown-toggle"
-            data-toggle="dropdown">
-            Sort by: <strong v-cloak>{{!sort}}</strong>
-            <span class="caret"></span>
-          </button>
-          <div class="dropdown-menu" role="menu">
-            <a class="dropdown-item" href v-on:click="changeSort('hostname', $event)">
-              <i v-bind:class="['fa fa-fw', {'fa-check': sort == 'hostname'}]"></i>
-              Hostname
-            </a>
-            <a class="dropdown-item" href v-on:click="changeSort('status', $event)">
-              <i v-bind:class="['fa fa-fw', {'fa-check': sort == 'status'}]"></i>
-              Status
-            </a>
-          </div>
-        </div>
-        <div class="dropdown">
-          <button type="button" class="btn btn-secondary dropdown-toggle"
-            data-toggle="dropdown">
-            Groups ({{! groupsFilter.length || 'all' }})
-            <span class="caret"></span>
-          </button>
-          <div class="dropdown-menu" role="menu">
-            <a class="dropdown-item" href="#" v-for="group in groups" v-on:click="toggleGroupFilter(group, $event)">
-              <i v-bind:class="['fa fa-fw', {'fa-check': groupsFilter.indexOf(group) != -1 }]"></i>
-              {{! group }}
-            </a>
-          </div>
-        </div>
-      </form>
-    </div>
-    <div class="col">
-      <p class="text-secondary text-right mt-2 mb-0 mr-4">Refreshed every 1m.</p>
-    </div>
-  </div>
-  <div class="row instance-list">
-    <div
-      v-for="instance, instanceIndex in filteredInstances"
-      :key="instance.hostname + instance.pg_port"
-      v-cloak
-      class="col-xs-12 col-sm-6 col-md-4 col-lg-3 col-xl-2 pb-3">
-      <instance-home-card
-        v-bind:instance="instance"
-        v-bind:status_value="getStatusValue(instance)"
-        v-bind:index="instanceIndex"
-      ></instance-home-card>
-    </div>
-    <p v-if="!loading && instances.length == 0" v-cloak>
-    {% try %}
-    {% if role.is_admin %}
-    No instance is available yet.<br>
-    Go to <strong>Settings</strong> to add or configure instances.
-    {% else %}
-    You don't have access to any instance.<br>
-    Please contact an administrator.
-    {% end %}
-    {% except %}
-    {% end %}
-    </p>
-  </div>
 </div>
 {% end %}
